fix: ensure Java interopt
The only issue found by this is that we were not able to read
Snappy-compressed data produced by Java clients, because Java uses its
own weird framing format. Funnily Java clients can read the Snappy data
what we produce. This is also in line with the rdkafka implemention
(they can consume the "normal" and the special "Java" format but only
produce the "normal" one).